# docs/openapi.yaml
openapi: 3.0.3
info:
    title: Tiles API
    description: MVP of the Tiles API
    version: 2.0.0

paths:
    /spatial-objects:
        get:
            summary: Get spatial objects
            description: Returns geometries from a BigQuery table as a GeoJSON FeatureCollection. Only the geometry column is queried; each feature contains the parsed geometry (any supported type).
            operationId: getSpatialObjects
            parameters:
                - name: tableFqn
                  in: query
                  required: true
                  description: Fully qualified name of the BigQuery table (e.g. project.dataset.table).
                  schema:
                      type: string
                      example: "my-project.my_dataset.my_geo_table"
                - name: geoColumn
                  in: query
                  required: false
                  description: Name of the geometry column in the table. Defaults to "geom".
                  schema:
                      type: string
                      default: "geom"
            responses:
                "200":
                    description: GeoJSON FeatureCollection of spatial objects.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GeoJSONFeatureCollection"
                            example:
                                type: "FeatureCollection"
                                features:
                                    - type: "Feature"
                                      geometry:
                                          type: "Point"
                                          coordinates: [100.0, 0.0]
                                    - type: "Feature"
                                      geometry:
                                          type: "LineString"
                                          coordinates:
                                              [
                                                  [0, 0],
                                                  [1, 1],
                                                  [2, 2],
                                              ]
                                    - type: "Feature"
                                      geometry:
                                          type: "Polygon"
                                          coordinates:
                                              [
                                                  [
                                                      [0, 0],
                                                      [1, 0],
                                                      [1, 1],
                                                      [0, 1],
                                                      [0, 0],
                                                  ],
                                              ]
                "400":
                    description: Bad request (validation errors for query parameters).
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorResponse"
                "500":
                    description: Server error (e.g. BigQuery or table access failure).

    /tilesets/{tilesetFQN}/tiles/{z}/{x}/{y}:
        get:
            summary: Get a tile from a tileset
            description: >
                Returns a single vector tile (MVT) from a BigQuery tileset table.
                The tileset table is expected to have columns `z`, `x`, `y`, and `data`,
                where `data` contains the tile in MVT format, gzipped and base64-encoded.
                The endpoint decodes and decompresses the tile before returning raw MVT binary.
            operationId: getTile
            parameters:
                - name: tilesetFQN
                  in: path
                  required: true
                  description: Fully qualified name of the BigQuery tileset table (e.g. project.dataset.tileset).
                  schema:
                      type: string
                      example: "my-project.my_dataset.my_tileset"
                - name: z
                  in: path
                  required: true
                  description: Zoom level of the tile.
                  schema:
                      type: integer
                      minimum: 0
                      example: 14
                - name: x
                  in: path
                  required: true
                  description: X coordinate (column) of the tile.
                  schema:
                      type: integer
                      minimum: 0
                      example: 4602
                - name: y
                  in: path
                  required: true
                  description: Y coordinate (row) of the tile.
                  schema:
                      type: integer
                      minimum: 0
                      example: 9819
            responses:
                "200":
                    description: The vector tile in MVT format.
                    content:
                        application/vnd.mapbox-vector-tile:
                            schema:
                                type: string
                                format: binary
                "400":
                    description: Bad request (validation errors for path parameters).
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorResponse"
                "404":
                    description: Tile not found at the specified z/x/y coordinates.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/NotFoundErrorResponse"
                "500":
                    description: Server error (e.g. BigQuery or table access failure).

components:
    schemas:
        ValidationFieldError:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: string
                    description: Machine-readable error code (e.g. required, invalid_format).
                message:
                    type: string
                    description: Human-readable error message.

        ValidationErrorResponse:
            type: object
            required:
                - errors
            properties:
                errors:
                    type: object
                    description: Map of field names to lists of validation errors.
                    additionalProperties:
                        type: array
                        items:
                            $ref: "#/components/schemas/ValidationFieldError"
                    example:
                        tableFqn:
                            - code: "required"
                              message: "tableFqn is required"

        NotFoundErrorResponse:
            type: object
            required:
                - error
            properties:
                error:
                    type: string
                    description: Human-readable error message.
                    example: "Tile not found at z=14, x=4602, y=9819"

        GeoJSONFeatureCollection:
            type: object
            required:
                - type
                - features
            properties:
                type:
                    type: string
                    enum: ["FeatureCollection"]
                features:
                    type: array
                    items:
                        $ref: "#/components/schemas/GeoJSONFeature"

        GeoJSONFeature:
            type: object
            required:
                - type
                - geometry
            properties:
                type:
                    type: string
                    enum: ["Feature"]
                geometry:
                    $ref: "#/components/schemas/GeoJSONGeometry"

        GeoJSONGeometry:
            oneOf:
                - $ref: "#/components/schemas/GeoJSONPoint"
                - $ref: "#/components/schemas/GeoJSONLineString"
                - $ref: "#/components/schemas/GeoJSONPolygon"
                - $ref: "#/components/schemas/GeoJSONMultiPoint"
                - $ref: "#/components/schemas/GeoJSONMultiLineString"
                - $ref: "#/components/schemas/GeoJSONMultiPolygon"
                - $ref: "#/components/schemas/GeoJSONGeometryCollection"
            discriminator:
                propertyName: type

        GeoJSONPoint:
            type: object
            required:
                - type
                - coordinates
            properties:
                type:
                    type: string
                    enum: ["Point"]
                coordinates:
                    type: array
                    minItems: 2
                    maxItems: 2
                    items:
                        type: number
                    description: "[longitude, latitude] in WGS84."

        GeoJSONLineString:
            type: object
            required:
                - type
                - coordinates
            properties:
                type:
                    type: string
                    enum: ["LineString"]
                coordinates:
                    type: array
                    items:
                        type: array
                        minItems: 2
                        maxItems: 2
                        items:
                            type: number
                    description: Array of [longitude, latitude] positions.

        GeoJSONPolygon:
            type: object
            required:
                - type
                - coordinates
            properties:
                type:
                    type: string
                    enum: ["Polygon"]
                coordinates:
                    type: array
                    items:
                        type: array
                        items:
                            type: array
                            minItems: 2
                            maxItems: 2
                            items:
                                type: number
                    description: Array of linear rings (first is exterior, rest are holes).

        GeoJSONMultiPoint:
            type: object
            required:
                - type
                - coordinates
            properties:
                type:
                    type: string
                    enum: ["MultiPoint"]
                coordinates:
                    type: array
                    items:
                        type: array
                        minItems: 2
                        maxItems: 2
                        items:
                            type: number
                    description: Array of [longitude, latitude] positions.

        GeoJSONMultiLineString:
            type: object
            required:
                - type
                - coordinates
            properties:
                type:
                    type: string
                    enum: ["MultiLineString"]
                coordinates:
                    type: array
                    items:
                        type: array
                        items:
                            type: array
                            minItems: 2
                            maxItems: 2
                            items:
                                type: number
                    description: Array of line strings.

        GeoJSONMultiPolygon:
            type: object
            required:
                - type
                - coordinates
            properties:
                type:
                    type: string
                    enum: ["MultiPolygon"]
                coordinates:
                    type: array
                    items:
                        type: array
                        items:
                            type: array
                            items:
                                type: array
                                minItems: 2
                                maxItems: 2
                                items:
                                    type: number
                    description: Array of polygons.

        GeoJSONGeometryCollection:
            type: object
            required:
                - type
                - geometries
            properties:
                type:
                    type: string
                    enum: ["GeometryCollection"]
                geometries:
                    type: array
                    items:
                        $ref: "#/components/schemas/GeoJSONGeometry"
                    description: Array of geometry objects.